+++ core/components/menukeeper/lexicon/en/setting.inc.php
@@ -5,5 +5,7 @@
  * @package menukeeper
  * @subpackage lexicon
  */
+$_lang['setting_menukeeper.cache_permissions'] = 'Cache Permission';
+$_lang['setting_menukeeper.cache_permissions_desc'] = 'Cache the permission of each menu entry.';
 $_lang['setting_menukeeper.debug'] = 'Debug';
 $_lang['setting_menukeeper.debug_desc'] = 'Log debug information in the MODX error log.';