+++ core/components/menukeeper/docs/changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-10-03
+
+### Changed
+
+- Cache the menu permissions too when menukeeper.cache_permissions is enabled
+
+### Fixed
+
+- Fix an undefined variable $menuObject on install
+
 ## [1.0.3] - 2024-06-25
 
 ### Changed